debug: implemented on rococo
